requirements.txt
bcrypt<4.0.0
fastapi[standard]>=0.112.0,<0.112.3
pydantic>=2.7.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2>=2.9.9,<3.0.0
load_dotenv>=0.1.0,<0.2.0
PyJWT>=1.7.1,<3.0.0
sqladmin[full]>=0.18.0,<1.0.0
fastapi-mail
passlib[bcrypt]
fastapi-pagination
alembic
Faker
apscheduler